% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openLab.R
\name{openLab}
\alias{openLab}
\title{Open Lab Function}
\usage{
openLab(name = "main")
}
\arguments{
\item{name}{Name of the lab to open. Defaults to 'intro'. Options include 'robots', 'simple', 'web_scraping', 'cosponsors', and 'api'.}
}
\description{
This function allows you to open the lab files.
}
\examples{
openLab('robots')
}
\keyword{labs}
